refactoring folder structure, decoupling app structure even more
