+++ src/js/app/index.js
@@ -6,11 +6,14 @@ import store from './store';
 import newApp from './app';
 import newRouter from './router';
 
-const Router = newRouter();
-const App = newApp({ Provider, store, Router });
 
-ReactDOM.render(
-  <div>{App}</div>
-  ,
-  document.getElementById('app'),
-);
+export function run(domElement) {
+  const Router = newRouter();
+  const App = newApp({ Provider, store, Router });
+
+  ReactDOM.render(
+    <div>{App}</div>
+    ,
+    document.getElementById(domElement),
+  );
+};
