+++ src/js/app/router.js
@@ -4,8 +4,8 @@ import {
   Switch,
 } from 'react-router-dom';
 
-import { routes as notFoundRoutes } from './notFound';
-import { routes as authRoutes } from './auth';
+import { routes as notFoundRoutes } from '../modules/notFound';
+import { routes as authRoutes } from '../modules/auth';
 
 export default function createRouter() {
   return (
