+++ src/js/index.js
@@ -1,0 +1,3 @@
+import * as App from './app';
+
+App.run('appContainer');
