+++ src/js/libs/api/constants.js
@@ -1,4 +1,2 @@
 export const API_HOST = process.env.API_HOST || '/';
 export const EMAIL_LOGIN_PATH = '/auth/login';
-
-export default API_HOST;
