+++ src/js/libs/jwt.js
