+++ src/js/modules/auth/components/EmailInput.js
