+++ src/js/modules/auth/components/PasswordInput.js
