+++ src/js/modules/auth/routes.js
@@ -4,7 +4,7 @@ import { Route } from 'react-router-dom';
 import { LoginView } from './components';
 
 
-export default function notFoundRoutes(basePath) {
+export default function authRoutes(basePath) {
   return (
     <Route path={path.join(basePath, '/login')} render={() => <LoginView />} />
   );
