+++ src/js/modules/game/index.js
